+++ src/main/java/com/k48/stock_management_system/services/impl/ClientServiceImpl.java
@@ -4,9 +4,11 @@ package com.k48.stock_management_system.services.impl;
 import com.k48.stock_management_system.dto.ClientDto;
 import com.k48.stock_management_system.exceptions.EntityNotFoundException;
 import com.k48.stock_management_system.exceptions.ErrorCode;
+import com.k48.stock_management_system.exceptions.InvalidOperationException;
 import com.k48.stock_management_system.model.Client;
+import com.k48.stock_management_system.model.CmdeClient;
 import com.k48.stock_management_system.repositories.ClientRepository;
-import com.k48.stock_management_system.services.AbstractService;
+import com.k48.stock_management_system.repositories.CmdeClientRepository;
 import com.k48.stock_management_system.services.ClientService;
 import com.k48.stock_management_system.validator.ObjectValidator;
 import lombok.RequiredArgsConstructor;
@@ -14,7 +16,6 @@ import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.Optional;
-import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
@@ -22,12 +23,13 @@ public class ClientServiceImpl implements ClientService {
 
     private final ObjectValidator<ClientDto> objectValidator;
     private final ClientRepository clientRepository;
+    private final CmdeClientRepository commandeClientRepository;
 
 
     @Override
     public ClientDto save(ClientDto clientDto) {
         objectValidator.validate(clientDto);
-        return ClientDto.toDto(clientRepository.save(ClientDto.toEntity(clientDto)));
+        return ClientDto.fromEntity(clientRepository.save(ClientDto.toEntity(clientDto)));
     }
 
     @Override
@@ -38,7 +40,7 @@ public class ClientServiceImpl implements ClientService {
         return
                 clientRepository
                         .findById(id)
-                        .map(ClientDto::toDto)
+                        .map(ClientDto::fromEntity)
                         .orElseThrow(
                                 ()-> new EntityNotFoundException("client with ID " + id + " not found"+ ErrorCode.ARTICLE_NOT_FOUND)
                         );
@@ -56,7 +58,7 @@ public class ClientServiceImpl implements ClientService {
                                 ()-> new EntityNotFoundException(" EMPTY LIST OF CLIENTS")
                         )
                         .stream()
-                        .map(ClientDto::toDto)
+                        .map(ClientDto::fromEntity)
                         .toList();
     }
 
@@ -66,8 +68,13 @@ public class ClientServiceImpl implements ClientService {
             return null;
         }
         Client client = clientRepository.findById(id).orElseThrow(()-> new EntityNotFoundException("client with ID " + id + " not found"));
+        List<CmdeClient> commandeClients = commandeClientRepository.findAllByClientId(id);
+        if (!commandeClients.isEmpty()) {
+            throw new InvalidOperationException("Impossible de supprimer un client qui a deja des commande clients",
+                    ErrorCode.CLIENT_ALREADY_IN_USE);
+        }
         clientRepository.delete(client);
 
-        return ClientDto.toDto(client);
+        return ClientDto.fromEntity(client);
     }
 }
